Extract s3 operations
